+++ src/metldata/builtin_transformations/delete_content_subschema/instruction.py
@@ -1,26 +1,15 @@
 # Copyright 2021 - 2024 Universität Tübingen, DKFZ, EMBL, and Universität zu Köln
 # for the German Human Genome-Phenome Archive (GHGA)
-#
+# 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-#
+# 
 #     http://www.apache.org/licenses/LICENSE-2.0
-#
+# 
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Models for instructions used in the 'delete content subschema' transformation."""
-
-from pydantic import Field
-from pydantic_settings import BaseSettings
-
-
-class DeleteSubschemaInstruction(BaseSettings):
-    """A model describing an instruction to remove a subschema for a class at a given content path."""
-
-    class_name: str = Field(..., description="The name of the class to modify.")
-    content_path: str = Field(..., description="Content schema property path.")
