+++ tests/fixtures/example_transformations/delete_content_subschema/multiple/transformed.schemapack.yaml
